pyproject.toml
[project]
name = "skeleton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cmaes>=0.12.0",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "matplotlib>=3.10.8",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "pytorch-lightning>=2.6.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
